all buffer lengths are of size_t now
git-svn-id: http://svn.miranda-ng.org/main/trunk@11173 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/IcqOscarJ/src/capabilities.h
@@ -36,8 +36,8 @@ typedef BYTE shortcapstr[BINARY_SHORT_CAP_SIZE];
 
 extern const capstr capShortCaps;
 
-capstr* MatchCapability(BYTE *buf, int bufsize, const capstr *cap, int capsize = BINARY_CAP_SIZE);
-capstr* MatchShortCapability(BYTE *buf, int bufsize, const shortcapstr *cap);
+capstr* MatchCapability(BYTE *buf, size_t bufsize, const capstr *cap, size_t capsize = BINARY_CAP_SIZE);
+capstr* MatchShortCapability(BYTE *buf, size_t bufsize, const shortcapstr *cap);
 
 
 #endif /* __CAPABILITIES_H */
