+++ protocols/IcqOscarJ/src/fam_0bstatus.cpp
@@ -25,7 +25,7 @@
 
 #include "icqoscar.h"
 
-void CIcqProto::handleStatusFam(unsigned char *pBuffer, WORD wBufferLength, snac_header* pSnacHeader)
+void CIcqProto::handleStatusFam(unsigned char *pBuffer, size_t wBufferLength, snac_header* pSnacHeader)
 {
 	switch (pSnacHeader->wSubtype) {
